+++ packages/components/ui/Accordion.tsx
@@ -1,8 +1,7 @@
 'use client'
+import { CheckButton, Icon } from '@/components'
 import clsx from 'clsx'
 import { useState } from 'react'
-import { CheckButton } from './CheckButton'
-import { Icon } from './Icon/Icon'
 
 interface AccordionProps {
   title: string
