+++ packages/components/ui/CheckButton.tsx
@@ -1,4 +1,4 @@
-import { Icon } from './Icon/Icon'
+import { Icon } from '@/components'
 
 interface CheckButtonProps {
   isChecked?: boolean
