+++ packages/tsconfig.json
@@ -1,7 +1,11 @@
 {
   "compilerOptions": {
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "jsx": "react-jsx",
+    "baseUrl": ".",
     "paths": {
-      "@/components/*": ["./*"]
+      "@/*": ["./*"]
     }
   }
 }
